Update MSTP
MSTP function was changed to output values in range [0, 1] instead of RGB [0 - 255] image. There was also a problem with the dtype of output, which was now changed from float64 to float32. (function was returning an array with 8bit values stored in float64 format).

TODO: save_to_8bit & change jupyter notebook with tutorial