NaNs handling for SVF
Converting nodata to NaNs moved from sky_view_factor_compute() to sky_view_factor(). Nodata has to be converted to NaNs before anything is done with the array.

Previously, nodata values were devided by resolution before it was converted to NaNs, which resulted in weird edge effects of the "NaN edges" of the dataset (i.e. not edges of the array itself, but where valid data meets nodata).