Slope np.roll nan values
When calculating derivatives, if there is Nan value in a rolled array, change it to the value from the original array. Because we are always rolling just by 1 pixel, this has the same effect as array padding with edge values (nearest neighbour). Array padding is only applied at the borders of an array, if we have border between valid data and Nan data, this new roll function applies the same effect.