Update tiled_multiprocess.py
The final result is now always a TIF file of the same size as source image. If it was running tiled processing, then vrt is cropped to the same size as source  (nodata at the edge clipped away) and joined into actual mosaic. If image is small enough, the whole image is processed as single image. The naming of the result file keeps the RVT naming conventions.