Fix raster shift when VRT to mosaic
The VRT extents may differ from the original file if nodata values are present at the edges. Previously, a shift (or stretch) occurred because a smaller raster was being forced into the original extents. This size difference arises from how tiles are handled—if an edge tile contains only nodata, it is ignored, resulting in a smaller output but only along the nodata edge.