Paused development
Use test_new_vis.ipynb for continued development. Currently we have created first step for the modified RRIM image, applying a custom colormap to SLRM visualization. Changes in tiled_multiprocess.py and _test_tiled_new_vis.py are NOT FINISHED, but the code is being prepared for inclusion of the new visualization combination.